Add lift combinators for HttpService. (#18)
* Add lift combinator for HttpService.

HttpService instances cannot be used where a `T: Service` is required.
`LiftService` is provided to adapt an `HttpService` instance to
`Service`.

* Add lift_ref to HttpService.

This is the same as `lift`, but operates on an HttpService reference.

Unfortunately, due to coherence, &mut T: HttpService cannot directly
implement either HttpService or Service. Thus, the reference must be
lifted in order to use it when either a T: HttpService or T: Service is
required.